+++ demos/nodes/testsmooth.gd
@@ -17,11 +17,15 @@ func _ready() -> void:
 	Engine.iterations_per_second = 5
 	
 	initial_states = {
+		# The 3D nodes
 		"nsmooth3": $s3d/non_smooth.global_transform,
 		"ysmooth3": $s3d/smoothed.global_transform,
+		"asmooth3": $s3d/autosmoothed.global_transform,
 		
+		# The 2D nodes
 		"nsmooth2": $s2d/non_smooth.global_transform,
 		"ysmooth2": $s2d/smoothed.global_transform,
+		"asmooth2": $s2d/autosmoothed.global_transform,
 	}
 	
 	# By default hide the "ghost" of smoothed objects
@@ -34,6 +38,7 @@ func _ready() -> void:
 	# tell the moment the reset has happened and give the change to perform the teleportation
 	# of the smoothed nodes
 	SharedUtils.connector($s2d/smoothed, "performed_reset", self, "on_rigid2d_reset", [$s2d/smoothed/Smooth2D, initial_states.ysmooth2])
+	SharedUtils.connector($s2d/autosmoothed, "performed_reset", self, "on_rigid2d_asmooth_reset", [$s2d/autosmoothed/AutoInterpolate, initial_states.asmooth2])
 	
 	
 	setup_hud()
@@ -56,21 +61,26 @@ func _on_bt_replay_pressed() -> void:
 	$s3d/non_smooth.angular_velocity = Vector3()
 	$s3d/smoothed.linear_velocity = Vector3()
 	$s3d/smoothed.angular_velocity = Vector3()
+	$s3d/autosmoothed.linear_velocity = Vector3()
+	$s3d/autosmoothed.angular_velocity = Vector3()
 	
 	$s3d/non_smooth.global_transform = initial_states.nsmooth3
 	$s3d/smoothed.global_transform = initial_states.ysmooth3
-	
+	$s3d/autosmoothed.global_transform = initial_states.asmooth3
 	
 	# Even though the "can_sleep" property is set to false, ensure the rigid bodies are not sleeping
 	$s3d/non_smooth.sleeping = false
 	$s3d/smoothed.sleeping = false
+	$s3d/autosmoothed.sleeping = false
 	
 	$s2d/non_smooth.reset_to(initial_states.nsmooth2)
 	$s2d/smoothed.reset_to(initial_states.ysmooth2)
+	$s2d/autosmoothed.reset_to(initial_states.asmooth2)
 	
 	# The smoothed object must be teleported
 	if ($hud/pnl/chk_useteleport.pressed):
 		$s3d/smoothed/Smooth3D.snap_to_target()
+		$s3d/AutoInterpolate.snap_to_target()
 		# The 2D objects will be "teleported" based on a signal, handled by the
 		# on_rigid2d_reset() function (bellow)
 
@@ -81,6 +91,10 @@ func on_rigid2d_reset(s: Smooth2D, t: Transform2D) -> void:
 		# Defer the call for one extra frame, to give the change of the target object to be moved
 		s.call_deferred("teleport_to", t)
 
+
+func on_rigid2d_asmooth_reset(s: AutoInterpolate, t: Transform2D) -> void:
+	if ($hud/pnl/chk_useteleport.pressed):
+		s.call_deferred("teleport_to", t)
 
 
 func setup_hud() -> void:
@@ -116,6 +130,7 @@ func _on_vsync_toggled(pressed: bool) -> void:
 func _on_showghost_toggled(pressed: bool) -> void:
 	$s3d/smoothed/refmesh.visible = pressed
 	$s2d/smoothed/refsprite.visible = pressed
+	$s2d/autosmoothed/refsprite.visible = pressed
 
 func _on_show3d_toggled(pressed: bool) -> void:
 	$s3d.visible = pressed